feat: output new release version if one made (#31)
## Pull Request Description

### Overview

This pull request introduces a new feature that outputs the new release
version when a release is created in GitHub Actions.

This is a feature that’s good to have, but I mostly implemented it to
have a integration test for pre-release versions.

### Changes Made

- **Feature Addition**: Added functionality to output the new release
version when a release is made in GitHub Actions. This helps users
easily identify the current version without needing to check additional
logs or files.
- **Testing**: Implemented tests to ensure that the pre-release version
is correctly created and asserted during the release process. This
ensures the reliability of the new output feature.

<!-- branch-stack -->

- `alpha`
  - \#31 :point\_left: