fix: github actions set-output is deprecated
When running the tool, I get warning:
Warning: The `set-output` command is deprecated and will be disabled soon. Please upgrade to using Environment Files. For more information see: https://github.blog/changelog/2022-10-11-github-actions-deprecating-save-state-and-set-output-commands/

But I'm using the latest 1.11.0 version of actions/core npm module so this shouldn't happen? I found the PR that implemented a fix for the deprecation, https://github.com/actions/toolkit/pull/1178, and in it, I see that it checks if it has access to the GITHUB_OUTPUT environment variable. If it doesn't, it will fallback to previous behavior.

commit-id:c8c672a4