refactor: move git commands out of action.yml and into deno code (#44)
## Problem
<!-- A clear description of the problem that this pull request is
solving. -->

Moved some git commands out of action.yml and into the tool's source
code.

- decouples the tool from the CI environment some more
- makes the code easier to understand as it's encapsulated
- better performance as we no longer need to pull *all* branches in a
repo, but just the ones we need.

## Solution
<!-- Describe the approach you took to solve the problem and the changes
made in this pull request. -->

Implement the git commands into the source code. 

## Testing
<!-- Choose one of the below options for how you tested the code change.
Include any specific setup or instructions for testing. -->
- [X] Added automated tests. 
- [X] Manually tested. If you check this box, provide instructions for
others to test, too.

Ran the tool on this PR 😄 

## Notes for reviewers 
<!-- If there is any additional information you would like to share with
the person reviewing this pull request, please provide it here. -->