fix: errors with tool not being able to import dependencies runtime
When the tool runs a deployment, it currently fails because when it tries to run the index.ts file remotely from GitHub, it then will eventually hit an import of a third-party dependency. And because we just switched to the lock file and deno.json file, the import statements are no longer pointing directly to a server that it can download from. It has to be able to read from the deno.json file. So there's a few options on how to fix this, but I think the future of this project is to go to compiled binaries anyway and leveraging lock files anyway. So moving in that direction here sounded like the easiest solution of creating a binary and then running that where it will find all the dependencies no matter what.