feat!: get latest release step is passed in from user
To make the tool as flexible as possible, the goal is that the user can define the behavior for every step. User can now define the code to run to get the latest release made when previously, it was hard-coded that you use github release for the release.

Note: You cannot pass in code to set the latest release. So this commit doesn't make this feature fully workable yet.