feat: create get-latest-release GitHub Releases step that projects can use
Now that the project is moving to a behavior where users provide all scripts to execute, it's time to pull out all of the hard-coded steps of the project. This is the start of that by pulling out the get-latest-release step code into a step that can be run by users in their projects.

I even created a test function based on what we suggest for users to do!