build: fix alpha releases not being marked as latest.
in last deployment, I noticed that the github release made was not marked as latest. this is bad since github repos advertise the latest release version on the homepage of the repo. I learned you can't mark a pre-release version as latest. For this repo, we want to always use latest even though we are still in alpha.