refactor: should run on other CI services now
change the runtime behavior to no longer only be compatible with github actions.

change deno permissions to allow access to all environment variables. because of the env-ci package we are using to try and be compatible with other environments, we need access to dozens of environment variables now. We might need to also give access to read and write to many more paths, but for now I want to leave it without.