Circleci project setup (#70)
## Related GitHub Issues
<!-- Link to any related GitHub issues that this pull request addresses
or closes. -->

## Problem
<!-- A clear description of the problem that this pull request is
solving. -->

We have allowed this program to be run on more than just one CI
provider, but it's not Tested on a different provider yet. This pull
request enables CircleCI to run the tool on this repo, so we can better
test the CLI capabilities of this tool.

## Solution
<!-- Describe the approach you took to solve the problem and the changes
made in this pull request. -->

## Testing
<!-- Choose one of the below options for how you tested the code change.
Include any specific setup or instructions for testing. -->
- [ ] Added automated tests. 
- [ ] Manually tested. If you check this box, provide instructions for
others to test, too.

## Notes for reviewers 
<!-- If there is any additional information you would like to share with
the person reviewing this pull request, please provide it here. -->

---------

Co-authored-by: circleci-app[bot] <127350680+circleci-app[bot]@users.noreply.github.com>