fix: hide github actions debug logs for non-github action environments
If you run the CLI not on GitHub Actions, you will see tons of logs. And that's because we always log debug logs because we're assuming you're on GitHub Actions. Added some new logic here to run code if you're in GitHub Actions and run other code if you're not.