fix: cannot get PR stack from github when running in circleci
When we run the tool in Circle CI, an error gets thrown saying that it can't Create the stack of pull requests successfully from the GitHub API response. I believe this is a issue with the environment variables available from GitHub Actions or CircleCI. and the data types not being converted successfully with the module that we use. Forcing the value to be a number instead of keeping it as is, I think, should fix this issue.