fix: avoid git error 'fatal: a branch named 'fix-simulated-merge-circle'' already exists
When you run this tool on a pull request in GitHub Actions, GitHub Actions will create a temporary branch that represents the pull request. So that's why we need to check out the branch that exists As the head of the pull request. But when you run the tool not from GitHub Actions like CircleCI, the Head branch is already checked out. Therefore, this error will be thrown when running in those environments.