feat: for performance, you can filter the list of branches that are parsed and provided as input to steps
I did some performance testing with a couple repositories, big and small, and found that the default behavior of Parsing every command in every branch of the repository for some projects indeed would take just way too long. So this is a feature that allows you to filter the branches for big projects to still be able to use this tool.