feat: for performance, you can specify max number of commits that are parsed and provided as input
I did some performance testing with a couple repositories, big and small, and found that the default behavior of Parsing every commit in every branch of the repository for some projects indeed would take just way too long. we added a feature that allowed you to filter the branches that we filter by, now this is an extra step to limit number of commits on each branch. I tested (swiftlang) main branch with 180,000 commits and it took 5 minutes to run. input scripts do not need that many commits.