build: simplify get-next-release script to be smaller in size
I want the scripts that are created for this project to be a reflection of what users of this tool will Need to do when they create them as well and Right now, the script is quite long, and so I think that just means that we could do a better job in the tool to make developers' lives easier. To get them less work.

When looking at this script, I did find little ways here and there to reduce the complexity of just the JavaScript code and so I refactored some of it. But then after that, I realized that a lot of the work I was doing was determining the next semantic version. And I was doing that because the library that we use to determine the next semantic version was not behaving like I expected it. After looking in the spec of semantic versioning version I realized the reason that the library was behaving as unexpected is because the spec is quite vague when it comes to pre-release versioning rules (the `-alpha.1` part at the end). So instead of fighting the lib library or making my own library I decided to go an alternative approach and instead just not use pre release versioning at all and instead use pre 1.0 because:
1. it is strictly enforced by the spec
2. the library does behave as I was expect to reduce the code in the script
3. also when I've used pre-release version strings on teams in the past, it's always caused confusion. So it just makes me think I don't know how much it's worth it.