refactor: move running deploy step script into step runner
There are design patterns that have already been established for running the other step scripts. So the fact that we have a separate block of code that is only for running the deploy step script doesn't make any sense. Make the code more maintainable by moving it into the step runner, but also allow me to continue following all the same design patterns that need to be followed for all running of step scripts.