fix: build url for github actions when error
there is a bug where if decaf fails and we give you a URL to the build in the PR comments, that URL is undefined for github actions. Turns out, some CI environments in the dependency we're using doesn't have a build url. So we had to make it optional but then implement github actions manually.