feat: run multiple simulated merges in test mode in 1 call
This commit is the start of a feature where you can now run multiple simulated merges in one call to the tool. Previously, if you wanted to run multiple simulated merges in test mode, you'd have to first call actions checkout to reset the Git environment, and then run the tool.

Now, you should be able to just run the tool with multiple simulated merges and merge types. We will run it once and process all the merge types.

---

This commit is not a complete implementation of the feature. It is mostly a test to make sure that this is even technically possible. So, even though we might not use this solution in the end, there is some code in here that's nice to have because it sets up the environment for us to be able to run multiple simulated merges in test mode.

So, yeah, just kind of a quick and dirty proof of concept, running it on the CI to see if it even works.