change strategy for creating the clone so we point to remote repo, not local clone repo
The current implementation for how we create an isolated git clone is to clone off of the local repo. And so then, whenever you perform a git fetch, it's going to try to fetch from the local repo and not from remote.

Yeah, I tried to get things to work on the CI and ran into some errors with it not working. So this should fix that, where we're actually creating a true clone that has the same git configuration as the local one.