refactor git module to not take in exec params and be locked to a directory
this is a refactor ive been wanting to do a long time. rather then having to pass around exec and directory params all over the place, just construct a new git instance with those params in the constructor and then only manage the git object. also safer because git object is immutable.