fix: escape the pull request template string
When you pass in the template string as a param, there's a good chance that it uses different kinds of quotes inside of it, which then messes with the CLI Bash command that we're running, where we're trying to pass the string into decaf. So, as a way to properly escape that string you pass in, we write to a temporary file and then just pass that file to decaf.